Moved vocabulary index to modules
